refactor(discovery): relocate the vuln assessment stage into a subpackage (adr-0018)
First stage relocation of Phase 6: move the vulnerability scanner, its CVE
data providers (NVD/local/CISA KEV), and the assess Stage into
internal/discovery/vuln, with depguard enforcing the inward direction.

- Kernel keeps the result types Vulnerability + DeviceVulnerabilities (they are
  DiscoveredDevice fields, so moving them would cycle); the vuln package aliases
  them so the moved scanner/provider code is otherwise unchanged.
- Engine drops its concrete *VulnerabilityScanner field for an injected Assessor
  port (SetAssessor); the composition root builds vuln.NewStage(scanner,
  engine.Registry(), engine.EventBus()).
- depguard rule discovery-stage-direction bans the kernel from importing the
  stage subpackage (stages depend on the kernel, never the reverse).
- internal/api re-points the ~8 scanner-type references to vuln.* (result types
  and the /api/v1 wire shapes are unchanged); vuln tests + export_test accessors
  moved with the package.

Behaviour-preserving: golden HTTP byte-identical, schema unchanged, full api +
discovery + vuln suites green, lint clean (incl. the new depguard rule).
