+++ internal/discovery/vulnerability_types.go
@@ -1,0 +1,43 @@
+package discovery
+
+// vulnerability_types.go holds the vulnerability RESULT types that are part of
+// the device aggregate. They stay in the kernel because DiscoveredDevice embeds
+// *DeviceVulnerabilities (devices_types.go) and the registry/device-copy code
+// reads them — moving them would force the kernel to import the vuln stage and
+// create an import cycle. The vuln stage (internal/discovery/vuln) — the scanner,
+// CVE providers, and the Assessor adapter — lives in its own package and aliases
+// these types (ADR-0018, Phase 6).
+
+import "time"
+
+// Vulnerability represents a CVE vulnerability for a device.
+type Vulnerability struct {
+	CVEID       string    `json:"cveId"`       // CVE-2023-12345
+	Description string    `json:"description"` // Brief description
+	Severity    string    `json:"severity"`    // CRITICAL, HIGH, MEDIUM, LOW
+	Score       float64   `json:"score"`       // CVSS score (0-10)
+	Published   time.Time `json:"published"`
+	Modified    time.Time `json:"modified"`
+	References  []string  `json:"references"`  // URLs to advisories
+	AffectedCPE string    `json:"affectedCpe"` // CPE string that matched
+
+	// CISA KEV (Known Exploited Vulnerabilities) enrichment fields
+	ActivelyExploited bool   `json:"activelyExploited,omitempty"` // In CISA KEV catalog
+	RansomwareRelated bool   `json:"ransomwareRelated,omitempty"` // Known ransomware use
+	RequiredAction    string `json:"requiredAction,omitempty"`    // CISA remediation guidance
+	DueDate           string `json:"dueDate,omitempty"`           // CISA remediation deadline
+	OriginalSeverity  string `json:"originalSeverity,omitempty"`  // Pre-KEV severity (if escalated)
+}
+
+// DeviceVulnerabilities holds vulnerability scan results for a device.
+type DeviceVulnerabilities struct {
+	DeviceIP        string          `json:"deviceIp"`
+	MAC             string          `json:"mac"`
+	Hostname        string          `json:"hostname"`
+	Vendor          string          `json:"vendor"`
+	Product         string          `json:"product"`
+	Version         string          `json:"version"`
+	Vulnerabilities []Vulnerability `json:"vulnerabilities"`
+	ScanTime        time.Time       `json:"scanTime"`
+	Error           string          `json:"error,omitempty"`
+}
